Change in nvchad configs

+++ nvchad/install-nvchad.fish
@@ -1,8 +1,12 @@
 #!/usr/bin/fish
 
 # Remove old nchad if possible
-rm -rf ~/.config/nvim
-rm -rf ~/.local/share/nvim
+echo "Do you want to remove old nvim config? [y/N]"
+read answer
+if test "$answer" = "y" -o "$answer" = "Y"
+    rm -rf ~/.config/nvim
+    rm -rf ~/.local/share/nvim
+end
 
 # Clone new nvchad
  git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1 && nvim
