+++ nvchad/remove-nvchad.fish
@@ -1,5 +1,9 @@
 #!/usr/bin/fish
 
 # Remove old nchad if possible
-rm -rf ~/.config/nvim
-rm -rf ~/.local/share/nvim
+echo "Do you want to remove old nvim config? [y/N]"
+read answer
+if test "$answer" = "y" -o "$answer" = "Y"
+    rm -rf ~/.config/nvim
+    rm -rf ~/.local/share/nvim
+end
